add http status and handle exceptions (category)

+++ src/main/java/com/dauphine/blogger/services/impl/CategoryServiceImpl.java
@@ -1,12 +1,13 @@
 package com.dauphine.blogger.services.impl;
 
 
+import com.dauphine.blogger.exceptions.CategoryNameAlreadyExists;
+import com.dauphine.blogger.exceptions.CategoryNotFoundByIdException;
 import com.dauphine.blogger.models.Category;
 import com.dauphine.blogger.repositories.CategoryRepository;
 import com.dauphine.blogger.services.CategoryService;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,9 +21,14 @@ public class CategoryServiceImpl implements CategoryService {
 
 
     @Override
-    public Category create(String name) {
-        Category category = new Category(name);
-        return repository.save(category);
+    public Category create(String name) throws CategoryNameAlreadyExists {
+        final boolean alreadyExists = repository.existsByName(name);
+        if (alreadyExists) {
+            throw new CategoryNameAlreadyExists(name);
+        } else {
+            Category category = new Category(name);
+            return repository.save(category);
+        }
     }
     @Override
     public List<Category> getAll() {
@@ -30,23 +36,35 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Category getById(UUID id) {
-        return repository.findById(id).orElse(null);
+    public List<Category> getAllByName(String name){
+        return repository.findAllByName(name);
     }
 
     @Override
-    public Category updateName(UUID id, String name) {
+    public Category getById(UUID id) throws CategoryNotFoundByIdException {
+        return repository.findById(id).orElseThrow(() ->new CategoryNotFoundByIdException(id));
+    }
+
+    @Override
+    public Category updateName(UUID id, String name) throws CategoryNotFoundByIdException, CategoryNameAlreadyExists {
+        final boolean alreadyExists = repository.existsByName(name);
+
         Category category = getById(id);
-        if (category == null) {
-            return null;
+        if (alreadyExists) {
+            throw new CategoryNameAlreadyExists(name);
         }
         category.setName(name);
         return repository.save(category);
     }
 
     @Override
-    public boolean deleteById(UUID id) {
-        repository.deleteById(id);
-        return true;
+    public boolean deleteById(UUID id) throws CategoryNotFoundByIdException {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        } else {
+            throw new CategoryNotFoundByIdException(id);
+        }
+
     }
 }
